+++ mrz-3-edition/tester.m
@@ -119,6 +119,8 @@ choleskySolve([
 ], [23 32 33 31]');  % exercise 4.8
 % exercise4_9();  % 1.0e-14 * ...
 % exercise4_10();  % see image
+% exercise4_11();  % see image
+% exercise4_12();  % see image
 
 %% chapter 5
 % algorithms