+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 - gauss inverter
 - cholesky linear system solver
 - jacobi, Gauss-Seidel, SOR methods
-- exercise 4.7, 4.8, 4.9, 4.10, 4.11 with image
+- exercise 4.7, 4.8, 4.9, 4.10, 4.11, 4.12 with image
 
 ### Fixed
 - && operator in Jacobi function